move plasma frequency tests into their own test file
